tool: move tcti and sapi context init functions
